Rename `current_murs` module to `current_cases`
