+++ webservices/tasks/legal_docs.py
@@ -8,7 +8,7 @@ from celery_once import QueueOnce
 
 from webservices import utils
 from webservices.legal_docs.advisory_opinions import load_advisory_opinions
-from webservices.legal_docs.current_murs import load_current_murs
+from webservices.legal_docs.current_cases import load_current_murs
 from webservices.rest import db
 from webservices.tasks import app
 from webservices.tasks.utils import get_app_name
